\name{const_hazard}
\alias{const_hazard}
\title{
Constant Hazard Function
}
\description{
Hazard function for Exponential Family Distributions
}
\usage{
const_hazard(r, lambda)
}
\arguments{
  \item{r}{
runlength
}
  \item{lambda}{
value of lambda
}
}
\author{
Zachary Zanussi
}
\note{
For use inside \code{\link{onlineCPD}}
}

\seealso{
\link{onlineCPD}
}

